add BBKEY_RBK command
also fix some formatting issues and make the heuristic for
restarting traces more obvious. The heuristic can interfere
with stream interpretation if it triggers at the wrong time.

This is particularly true of routines involving the security key
because it looks like the JTAG driver is a different piece of
code for security keys inside the Vivado tools, based on how
drastically the timing idioms change between the "normal"
mode and the security key modes.